refactor(xod-core, xod-client): move functions for handling input in inspector widget to xod-client
