+++ packages/xod-client/src/editor/components/inspectorWidgets/Widget.jsx
@@ -1,6 +1,6 @@
 import R from 'ramda';
 import React from 'react';
-import { PROPERTY_TYPE_PARSE, PROPERTY_TYPE_MASK } from 'xod-core';
+import { PROPERTY_TYPE_PARSE, PROPERTY_TYPE_MASK } from '../../../utils/inputFormatting';
 
 import { KEYCODE } from '../../../utils/constants';
 import { noop } from '../../../utils/ramda';
