+++ packages/xod-client/src/utils/inputFormatting.js
@@ -1,0 +1,34 @@
+import R from 'ramda';
+import { PROPERTY_TYPE } from 'xod-core';
+
+const removeAllDotsExceptFirst = str =>
+  str.replace(/^([^.]*\.)(.*)$/, (a, b, c) => b + c.replace(/\./g, ''));
+
+/**
+ * transform value when input is in progress
+ */
+export const PROPERTY_TYPE_MASK = {
+  [PROPERTY_TYPE.BOOL]: R.identity,
+  [PROPERTY_TYPE.NUMBER]: R.compose(
+    R.when(
+      R.compose(R.equals('.'), R.head),
+      R.concat('0')
+    ),
+    removeAllDotsExceptFirst,
+    R.replace(/[^0-9.]/g, ''),
+    R.toString
+  ),
+  [PROPERTY_TYPE.STRING]: R.identity,
+  [PROPERTY_TYPE.PULSE]: R.identity,
+};
+
+export const PROPERTY_TYPE_PARSE = {
+  [PROPERTY_TYPE.BOOL]: v => !!v,
+  [PROPERTY_TYPE.NUMBER]: (v) => {
+    const float = parseFloat(v, 10);
+    // TODO: danger: return type is still Number | String
+    return isNaN(float) ? '' : float;
+  },
+  [PROPERTY_TYPE.STRING]: v => String(v),
+  [PROPERTY_TYPE.PULSE]: v => !!v,
+};
